[COMMON] thermal: fix prevent error
CID: 202888 202205 202201 201542 200628 200625

Resolved migration conflicts from kernel 4.9 to 4.14.
- rename cpufreq_device -> cpufreq_cdev
- rename gpufreq_device -> gpufreq_cdev

+++ drivers/thermal/cpu_cooling.c
@@ -383,7 +383,7 @@ static int lookup_static_power(struct cpufreq_cooling_device *cpufreq_cdev,
 	temperature  = temperature / 1000;
 
 	for (volt_index = 0; volt_index <= cpufreq_cdev->var_volt_size; volt_index++) {
-		if (voltage < cpufreq_cdev->var_table[volt_index * (cpufreq_cdev->var_temp_size + 1)]) {
+		if (voltage < cpufreq_cdev->var_table[volt_index * ((int)cpufreq_cdev->var_temp_size + 1)]) {
 			volt_index = volt_index - 1;
 			break;
 		}
