+++ drivers/thermal/isp_cooling.c
@@ -262,7 +262,7 @@ static int isp_apply_cooling(struct isp_cooling_device *isp_device,
 	if (isp_device->isp_state == cooling_state)
 		return 0;
 
-	isp_device->isp_state = cooling_state;
+	isp_device->isp_state = (unsigned int)cooling_state;
 
 	blocking_notifier_call_chain(&isp_notifier, ISP_THROTTLING, &cooling_state);
